% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_jstage_to_excel.R
\name{write_jstage_to_excel}
\alias{write_jstage_to_excel}
\title{Write J-Stage Data to Excel File}
\usage{
write_jstage_to_excel(data, file_name)
}
\arguments{
\item{data}{A list containing the metadata and entry data frames obtained from J-Stage.}

\item{file_name}{A character string specifying the name of the Excel file to save the data.}
}
\value{
None. The function writes data to an Excel file.
}
\description{
This function writes the J-Stage search results to an Excel file.
}
