% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jstage_metadata.R
\name{jstage_metadata}
\alias{jstage_metadata}
\title{Scrape J-STAGE Article Metadata}
\usage{
jstage_metadata(url, bibtex_file_name = "")
}
\arguments{
\item{url}{The URL of the J-STAGE article.}

\item{bibtex_file_name}{The file name to save the BibTeX entry. If empty, no file is saved.}
}
\value{
A list containing the article metadata.
}
\description{
Scrapes all available metadata from a J-STAGE article.
}
