% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jstage_references.R
\name{jstage_references}
\alias{jstage_references}
\title{Scrape J-STAGE References}
\usage{
jstage_references(url, depth = 1, quiet = TRUE)
}
\arguments{
\item{url}{The URL or DOI of the J-STAGE article web page.}

\item{depth}{Integer. The depth to which references should be scraped. Default is 1.}

\item{quiet}{logical. Suppress information regarding the progress.}
}
\value{
A data frame with the DOI of each reference.
}
\description{
Scrapes all available references from a J-STAGE article web page.
}
